+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Change Log
 
+## 0.0.2
+- Allows any LanguageTool supported language to be used through the use of supplemental extensions
+- Removes built-in English support
+
 ## 0.0.1
 - Initial release